requirements.txt
fastapi
uvicorn
faiss-cpu
cohere
PyMuPDF
python-docx
dotenv
